Brings over: oracle assertions, human_principal, runit migration,
tched_router rebrand, OVS durable flows, op-dbus-cli, cognitive-mcp
fan-in, and all accumulated fixes from the PR #30 line.

Identity/genesis files kept from cognitive-boot (the intended base).
lovable submodule removed (retired).

+++ crates/op-cognitive-mcp/src/rag_pipeline.rs
@@ -788,10 +788,7 @@ fn is_already_exists(err: &QdrantError) -> bool {
     match err {
         QdrantError::ResponseError { status }
         | QdrantError::ResourceExhaustedError { status, .. } => {
-            // qdrant-client currently carries Tonic 0.14 while the OP-DBUS
-            // workspace surface remains on 0.12. Their `Code` enums are
-            // distinct Rust types but share the canonical gRPC wire values.
-            status.code() as i32 == tonic::Code::AlreadyExists as i32
+            status.code() == tonic::Code::AlreadyExists
         }
         _ => false,
     }
